+++ web/apps/photos/src/components/AuthenticateUserModal.tsx
@@ -1,8 +1,8 @@
 import { checkSessionValidity } from "@/accounts/services/session";
 import log from "@/base/log";
 import { AppContext } from "@/new/photos/types/context";
+import type { DialogBoxAttributesV2 } from "@ente/shared/components/DialogBoxV2";
 import DialogBoxV2 from "@ente/shared/components/DialogBoxV2";
-import type { DialogBoxAttributesV2 } from "@ente/shared/components/DialogBoxV2/types";
 import VerifyMasterPasswordForm, {
     type VerifyMasterPasswordFormProps,
 } from "@ente/shared/components/VerifyMasterPasswordForm";
